+++ source/proxy_connection.c
@@ -466,7 +466,7 @@ static struct aws_http_message *s_build_h1_proxy_connect_request(struct aws_http
     }
 
     char port_str[20] = "\0";
-    snprintf(port_str, sizeof(port_str), "%d", (int)user_data->original_port);
+    snprintf(port_str, sizeof(port_str), "%u", user_data->original_port);
     struct aws_byte_cursor port_cursor = aws_byte_cursor_from_c_str(port_str);
     if (aws_byte_buf_append(&path_buffer, &port_cursor)) {
         goto on_error;
